Add support for hetero-nn pipeline Signed-off-by: weijingchen <talkingwallace@sohu.com>
